+++ sic/index.php
@@ -11,7 +11,7 @@ ini_set('display_errors', 'On');
 error_reporting(E_ALL);
 
 
-$siclight_version = "1.6.2";
+$siclight_version = "1.7";
 
 ?>
 <!DOCTYPE html>
@@ -50,11 +50,13 @@ $siclight_version = "1.6.2";
                     $dl_sum_class = 'hide';
                 }
 
-                // refresh all button markup
+                // refresh all / refresh selected button markup
                 if($ActiveSites['count']>0){
-                    $refresh_all_btn = "<button class='refresh-all uk-button uk-button-danger' type='button' uk-tooltip title='Refresh all active sites'><span uk-icon='icon: refresh'></span></button>\n";
+                    $refresh_all_btn = "<button class='refresh-all uk-button uk-button-danger' type='button' uk-tooltip title='Refresh all active sites'><span uk-icon='icon: refresh'></span> ALL</button>\n";
+                    $refresh_selected_btn = "<button class='refresh-selected uk-button uk-button-primary' type='button' uk-tooltip title='Refresh filtered sites'><span uk-icon='icon: refresh'></span> FILTERED</button>\n";
                 } else {
                     $refresh_all_btn = "";
+                    $refresh_selected_btn = "";
                 }
 
 
@@ -62,10 +64,13 @@ $siclight_version = "1.6.2";
                 <div class='uk-card uk-card-default'>
                     <div class='uk-card-header'>
                         <div uk-grid class='uk-child-width-expand'>
-                            <div><h2 class='uk-card-title'>Active Sites <span class='uk-badge'>".$ActiveSites['count']."</span></h2></div>
+                            <div>
+                                <h2 class='uk-card-title'>Active Sites <span class='uk-badge'>".$ActiveSites['count']."</span></h2>
+                            </div>
                             <div class='header-left'>
+                                {$refresh_selected_btn}
                                 {$refresh_all_btn}
-                                <a href='history/_summary-latest.csv' target='blank' class='uk-button uk-button-default download-summary {$dl_sum_class}' uk-tooltip title='Download latest summary CSV'><span uk-icon='icon: download'></span></a>
+                                <a href='history/_summary-latest.csv' target='blank' class='uk-button uk-button-default download-summary {$dl_sum_class}' uk-tooltip title='Download latest summary CSV'><span uk-icon='icon: download'></span> CSV</a>
                             </div>
                         </div>
                     </div>
@@ -92,9 +97,8 @@ $siclight_version = "1.6.2";
     <script src="js/jquery.ajaxq.js"></script>
     <script src="js/uikit.min.js"></script>
     <script src="js/uikit-icons.min.js"></script>
+    <script src="js/datatables.min.js"></script>
     <script src="js/main.js"></script>
-
-
-
+    
 </body>
 </html>